mybatis generator 自动生成实体类、接口和mapper文件

+++ member/src/main/java/com/example/member/controller/MemberConterller.java
@@ -17,7 +17,7 @@ public class MemberConterller {
     @Resource
     private MemberService memberService;
     @GetMapping("/count")
-    public int hello(){
+    public long hello(){
         return memberService.getMemberCount();
     }
 }
